Fix warnings

+++ src/crypto/keccak.hpp
@@ -31,7 +31,7 @@ namespace morum::crypto {
       };
       uint64_t(*a)[5] = state;
       uint8_t r = 1;  // LFSR
-      for (int i = 0; i < NUM_ROUNDS; i++) {
+      for (int i = 0; i < static_cast<int>(NUM_ROUNDS); i++) {
         // Theta step
         uint64_t c[5] = {};
         for (int x = 0; x < 5; x++) {
@@ -80,7 +80,7 @@ namespace morum::crypto {
         absorb();
       }
       // Uint64 array to bytes in little endian
-      for (int i = 0; i < HASH_LEN; i++) {
+      for (int i = 0; i < static_cast<int>(HASH_LEN); i++) {
         int j = i >> 3;
         hash[i] = static_cast<uint8_t>(state[j % 5][j / 5] >> ((i & 7) << 3));
       }
